locust/ecommerce/requirements.txt
-r ../requirements.txt
PyJWT==0.4.3
